+++ NEWS.md
@@ -2,6 +2,7 @@ develop
 ==============
 
 1. add Schoenfeld residual output for Cox models
+2. check for negative curvature before computing CIs
 
 Cyclops v3.5.0
 ==============